03_Source_Code/agency_banking/requirements.txt
# Wekeza Agency Banking System Requirements

# Core dependencies
streamlit>=1.28.0
mysql-connector-python>=8.0.33
pandas>=1.5.0

# Security and authentication
cryptography>=3.4.8
pyjwt>=2.4.0

# Data processing
numpy>=1.21.0
python-dateutil>=2.8.2

# API and web services
requests>=2.28.0
fastapi>=0.68.0
uvicorn>=0.15.0

# Utilities
python-dotenv>=0.19.0
pydantic>=1.8.0

# Development and testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Optional: For advanced features
# geopy>=2.2.0          # For geo-location calculations
# qrcode>=7.3.0         # For QR code generation
# pillow>=8.3.0         # For image processing
# reportlab>=3.6.0      # For PDF receipt generation